+++ crates/ironcalc-base/src/expressions/parser/static_analysis.rs
@@ -899,6 +899,7 @@ fn get_function_args_signature(kind: &Function, arg_count: usize) -> Vec<Signatu
         Function::Sequence => args_signature_scalars(arg_count, 1, 3),
         Function::Randarray => args_signature_scalars(arg_count, 0, 5),
         Function::Frequency => vec![Signature::Vector; arg_count],
+        Function::Textsplit => vec![Signature::Scalar; arg_count],
         Function::Sort => {
             let mut result = vec![Signature::Vector; arg_count];
             for item in result.iter_mut().skip(1) {
@@ -1551,6 +1552,7 @@ fn static_analysis_on_function(kind: &Function, args: &[Node]) -> StaticResult {
         Function::Rand => not_implemented(args),
         Function::Randarray => not_implemented(args),
         Function::Frequency => not_implemented(args),
+        Function::Textsplit => not_implemented(args),
         Function::Randbetween => scalar_arguments(args),
         Function::Eomonth => scalar_arguments(args),
         Function::Formulatext => not_implemented(args),