+++ crates/ironcalc-base/src/test/mod.rs
@@ -51,6 +51,7 @@ mod test_fn_minverse;
 mod test_fn_munit;
 mod test_fn_randarray;
 mod test_fn_frequency;
+mod test_fn_textsplit;
 mod test_fn_transpose;
 mod test_fn_bahttext;
 mod test_fn_percentile;